chatbot/requirements.txt
# Framework Web
flask==3.0.2
flask-cors==4.0.0
flask-sqlalchemy==3.1.1

# Configuração e Variáveis de Ambiente
python-dotenv==1.0.0

# Requisições HTTP e OAuth
requests==2.31.0

# Servidor de Produção
gunicorn==21.2.0

# Autenticação JWT
PyJWT>=2.8.0

# Utilitários do Flask
Werkzeug>=2.3.0

# Dependências do SQLAlchemy
SQLAlchemy>=2.0.0

# Fuzzy matching usado na base de conhecimento
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0

# Validação de dados e modelos
pydantic>=2.0.0

# Driver MySQL
pymysql>=1.1.0
cryptography>=41.0.0

# Testes
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-flask>=1.3.0